Entry 10 In Progress
Was able to incorporate JavaScript button but still not able to toggle it off and have the words revert back on another click.

Would like this button click to trigger the dropping of the postcard.

How do I resize images!!!! and if I do this with flexbox, how do I stack/overlap elements - the card on the line?