Entry 10
Worked out toggling the button between image and text.

Still need it to trigger the dropping of the card. It does seem to trigger a css class but on the second click, probably a result of if, else.

Added a larger letter at the bottom. Would like this to come out when you click on a postcard. this should be text embedded in the html but how will i have it fall?

Running into issues with certain elements not showing up in a wider media query as they are hidden in smaller ones. I'm using the same code as last entry but not sure why it's not working.